Refactor data pipeline and search engine for improved modularity
