+++ IMPLEMENTATION_SUMMARY.md
@@ -1,4 +1,4 @@
-# 🎯 **Implementation Summary: Production-Ready Microservice**
+# 🎯 **Implementation Summary**
 
 ## ✅ **Implementation Complete**
