+++ REFACTORING_SUMMARY.md
@@ -1,4 +1,4 @@
-# 🔧 **ACTUAL Code Refactoring Summary**
+# 🔧 **Code Refactoring Summary**
 
 ## ✅ **Real Refactoring Completed**
